refactor/transfers: move history to own mod
