+++ src/client_handler/transfers/mod.rs
@@ -6,6 +6,7 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
+mod history;
 mod replica;
 
 pub type Identity = PublicKey;
